Split up package.json - apps/action-server
Change-type: patch
