+++ apps/action-server/lib/worker.js
@@ -6,14 +6,15 @@
 
 const logger = require('@balena/jellyfish-logger').getLogger(__filename)
 const uuid = require('@balena/jellyfish-uuid')
-const packageJSON = require('../../package.json')
+const packageJSON = require('../../../package.json')
 const bootstrap = require('./bootstrap')
+const environment = require('@balena/jellyfish-environment')
 
 const DEFAULT_ERROR_CONTEXT = {
-	id: `TICK-ERROR-${packageJSON.version}`
+	id: `WORKER-ERROR-${packageJSON.version}`
 }
 
-const onError = (serverContext, error, message = 'Tick worker error') => {
+const onError = (serverContext, error, message = 'Worker error') => {
 	logger.exception(serverContext, message, error)
 	setTimeout(() => {
 		process.exit(1)
@@ -21,20 +22,21 @@ const onError = (serverContext, error, message = 'Tick worker error') => {
 }
 
 process.on('unhandledRejection', (error) => {
-	return onError(DEFAULT_ERROR_CONTEXT, error, 'Unhandled Tick Error')
+	return onError(DEFAULT_ERROR_CONTEXT, error, 'Unhandled Worker Error')
 })
 
 const startDate = new Date()
 uuid.random().then((id) => {
 	const context = {
-		id: `TICK-${packageJSON.version}-${id}`
+		id: `WORKER-${packageJSON.version}-${id}`
 	}
 
-	logger.info(context, 'Starting tick worker', {
+	logger.info(context, 'Starting worker', {
 		time: startDate.getTime()
 	})
 
-	return bootstrap.tick(context, {
+	return bootstrap.worker(context, {
+		metricsPort: environment.metrics.ports.app,
 		onError: (serverContext, error) => {
 			return onError(serverContext, error)
 		}
@@ -49,7 +51,7 @@ uuid.random().then((id) => {
 		const endDate = new Date()
 		const timeToStart = endDate.getTime() - startDate.getTime()
 
-		logger.info(context, 'Tick worker started', {
+		logger.info(context, 'Worker started', {
 			time: timeToStart
 		})
 	}).catch((error) => {
