+++ package.json
@@ -14,7 +14,7 @@
   },
   "deplint": {
     "files": [
-      "apps/action-server/**/*.js",
+      "apps/action-server/lib/**/*.js",
       "apps/server/*.js",
       "apps/server/default-cards/**/*.js",
       "apps/server/graphql/**/*.js",
@@ -62,6 +62,13 @@
     "@babel/preset-react": "^7.12.1",
     "@babel/register": "^7.12.1",
     "@balena/ci-task-runner": "^0.2.36",
+    "@balena/jellyfish-action-library": "^5.0.55",
+    "@balena/jellyfish-core": "^0.1.74",
+    "@balena/jellyfish-environment": "^2.2.39",
+    "@balena/jellyfish-queue": "^0.0.157",
+    "@balena/jellyfish-sync": "^3.0.34",
+    "@balena/jellyfish-uuid": "^1.0.16",
+    "@balena/jellyfish-worker": "^1.0.134",
     "ava": "^3.13.0",
     "babel-loader": "^8.1.0",
     "browser-env": "^3.3.0",
@@ -108,18 +115,9 @@
   },
   "dependencies": {
     "@babel/polyfill": "^7.12.1",
-    "@balena/jellyfish-action-library": "^5.0.55",
     "@balena/jellyfish-chat-widget": "^2.0.53",
     "@balena/jellyfish-client-sdk": "^2.2.13",
-    "@balena/jellyfish-core": "^0.1.74",
-    "@balena/jellyfish-environment": "^2.2.39",
-    "@balena/jellyfish-logger": "^0.0.114",
-    "@balena/jellyfish-metrics": "^0.0.130",
-    "@balena/jellyfish-queue": "^0.0.157",
-    "@balena/jellyfish-sync": "^3.0.34",
     "@balena/jellyfish-ui-components": "^2.0.46",
-    "@balena/jellyfish-uuid": "^1.0.16",
-    "@balena/jellyfish-worker": "^1.0.134",
     "@emotion/cache": "^10.0.29",
     "@emotion/core": "^10.0.35",
     "@octokit/plugin-retry": "^3.0.3",
