+++ test/containers/Dockerfile.test
@@ -7,7 +7,9 @@ ARG NPM_TOKEN
 COPY . ./
 RUN echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > ~/.npmrc && \
 	PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true npm ci --unsafe-perm && \
-	cd apps/server && npm ci && rm -f ~/.npmrc
+	cd /usr/src/jellyfish/apps/server && npm ci && \
+	cd /usr/src/jellyfish/apps/action-server && npm ci && \
+	rm -f ~/.npmrc
 
 # Copy in source and run catch-uncommitted, lint, and unit tests
 RUN make docs/assets/architecture.png ARCHITECTURE.md docker-compose.yml && \
